xengine_to_uvfits.py: workaround CASA bug regarding order of baselines
When doing concat(), CASA will swap the antennas in the baselines
(j,k) with j > k in all the files except the first one. This gives
all sorts of problems with plotms().

By swapping the baselines before writing to UVFITS to ensure that
j <= k, we prevent CASA from swapping them later.